requirements.txt
newsapi-python>=0.2.7
spacy>=3.7.7
python-dotenv>=1.0.0
transformers>=4.36.2
torch>=2.0.0
pytest>=7.4.3
requests>=2.31.0
pandas>=2.1.4